Implement the new parsing functions
This simplifies the parsing generator a great deal by adding an index to
our path data type. Path is now indexed by its length and the length of
the path must match the number of components of the url.

The project files are not split between examples and source.

The readme has been update accordingly with installation instructions
and information about how to run the example.